requirements.txt
openai ~= 1.61.0
google-generativeai ~= 0.8.4
pytubefix ~= 8.12.1
moviepy ~= 2.1.2